+++ crates/wdl-lint/src/rules/double_quotes.rs
@@ -3,6 +3,7 @@
 use wdl_ast::span_of;
 use wdl_ast::v1::Expr;
 use wdl_ast::v1::LiteralExpr;
+use wdl_ast::v1::LiteralStringKind;
 use wdl_ast::Diagnostic;
 use wdl_ast::Diagnostics;
 use wdl_ast::Document;
@@ -74,7 +75,7 @@ impl Visitor for DoubleQuotesRule {
         }
 
         if let Expr::Literal(LiteralExpr::String(s)) = expr {
-            if s.quote() != '"' {
+            if s.kind() == LiteralStringKind::SingleQuoted {
                 state.add(use_double_quotes(span_of(s)));
             }
         }